Make Prescription.quantity_float a FloatField instead of an IntegerField.
Former-commit-id: e6b8ab97943ace8a002267a62e122a6cb410bda0

+++ src/ESP/emr/models.py
@@ -781,7 +781,7 @@ class Prescription(BasePatientRecord):
     # This really should be quantity_string instead of quantity; but I don't 
     # want to break a bunch of other stuff right now.
     quantity = models.CharField(max_length=200, blank=True, null=True)
-    quantity_float = models.IntegerField(blank=True, null=True, db_index=True)
+    quantity_float = models.FloatField(blank=True, null=True, db_index=True)
     refills = models.CharField(max_length=200, blank=True, null=True)
     route = models.CharField(max_length=200, blank=True, null=True)
     status = models.CharField('Order Status', max_length=20, blank=True, null=True)
